% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPprior.R
\name{create_plot}
\alias{create_plot}
\title{Title}
\usage{
create_plot(b, X, Y)
}
\arguments{
\item{b}{scalar input should be the row number user wants to remove(has to be greater than 2)}

\item{X}{matrix input}

\item{Y}{vector input}
}
\description{
Title
}
