% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPprior.R
\name{gp_posterior}
\alias{gp_posterior}
\title{Title}
\usage{
gp_posterior(
  X,
  Y,
  b,
  col_num,
  sigma_hat,
  degree = NULL,
  choice = 1,
  l = NULL,
  alpha = NULL
)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
