% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPprior.R
\name{gp_prior}
\alias{gp_prior}
\title{GP Priors}
\usage{
gp_prior(
  X,
  Y,
  b,
  col_num,
  sigma_hat,
  degree = NULL,
  choice = 1,
  l = NULL,
  alpha = NULL
)
}
\arguments{
\item{X}{matrix input}

\item{Y}{vector input}

\item{b}{scalar input, should be the discontinuity point in the data}

\item{col_num}{scalar input allowing user to select particular column of X}

\item{sigma_hat}{vector input}

\item{degree}{scalar input, default is NULL if no argument given}

\item{choice}{scalar input 1 or 2, default is 1}

\item{l}{scalar input}

\item{alpha}{vector input}
}
\value{
mean function and covariance function
}
\description{
GP Priors
}
\examples{
X <- matrix(rnorm(24), nrow = 4)
y <- c(1, 2, 3, 4)
gp_prior(X, y, b = 3, sigma_hat = 0.3, l = 2)
}
