% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gp.R
\name{gp_prior}
\alias{gp_prior}
\title{GP Priors}
\usage{
gp_prior(
  Xc,
  Xt,
  Yc,
  Yt,
  sigma_hat,
  choice = 1,
  l = NULL,
  alpha = NULL,
  degree = NULL
)
}
\arguments{
\item{Xc}{vector input for the control group}

\item{Xt}{vector input for the test group}

\item{Yc}{vector input for the response of the control group}

\item{Yt}{vector input for the response of the test group}

\item{sigma_hat}{scalar input for covariance kernels}

\item{choice}{scalar input 1 or 2, default is 1, which is the squared covariance kernel. 2 is rational quadratic kernel}

\item{l}{scalar input for covariance kernel}

\item{alpha}{scalar input for rational quadratic kernel}

\item{degree}{scalar input, default is NULL if no argument given}
}
\value{
A list with the elements
\item{mean_control}{A nc length vector for the prior mean function for the control group}
\item{mean_treatment}{A nt length vector for the prior mean function for the treatment group}
\item{cov_c}{An nc by nc matrix for the prior covariance kernel for control group }
\item{cov_t}{An nt by nt matrix for the prior covariance kernel for treatment group}
}
\description{
GP Priors
}
\examples{
set.seed(100)
xc = seq(0, 1, length.out = 100)
xt = seq(1, 2, length.out = 100)
yc = 1 * xc + rnorm(100, 0, 0.25)
yt = 2 + 1*xt + rnorm(100, 0, 0.25)
gp_prior(xc, xt, yc, yt, sigma_hat = 1.2, choice = 1, l = 0.7)
gp_prior(xc, xt, yc, yt, sigma_hat = 1.2, choice = 2, l = 0.8, alpha = 2)
}
