% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPprior.R
\name{gp_prior}
\alias{gp_prior}
\title{GP Priors}
\usage{
gp_prior(degree = 2, X, y)
}
\arguments{
\item{degree}{scalar input, default is 2 if no argument given}

\item{X}{matrix input}

\item{y}{vector input}
}
\value{
multivariate normal distribution with given mean function and covariance function
}
\description{
GP Priors
}
\examples{
gp_prior(degree = 3, X, y)
}
